Refactor events results to a component, use component in topic content type pages
