+++ src/components/event-results/index.ts
@@ -1,0 +1,3 @@
+import EventResults from './event-results';
+
+export default EventResults;
